requirements.txt
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core==1.3.0
langchain-groq==1.1.2
langchain-huggingface==1.2.2
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.2
openai==2.32.0
pypdf==6.10.2
python-dotenv==1.2.2
qdrant-client==1.17.1
rank-bm25==0.2.2
sentence-transformers==5.4.1
streamlit==1.56.0
wikipedia-api==0.8.1
